+++ mod-openvino/CMakeLists.txt
@@ -43,7 +43,7 @@ set( NOISE_SUPPRESSION_SOURCES
 include_directories(noise_suppression)
 include_directories(noise_suppression/deepfilternet)
 
-set( AUDIO_SR_SOURCES 
+set( AUDIO_SR_SOURCES
      OVAudioSR.cpp
      OVAudioSR.h
      audio_sr/audiosr_common.h
@@ -73,6 +73,8 @@ set( SOURCES
       ${NOISE_SUPPRESSION_SOURCES}
       ${AUDIO_SR_SOURCES}
       ${DEMIX_SOURCES}
+      OVDemixerEffect.cpp
+      OVDemixerEffect.h
       OVMusicSeparation.cpp
       OVMusicSeparation.h
       OVMusicGenerationLLM.cpp
