+++ mod-openvino/OVMusicSeparation.cpp
@@ -39,128 +39,10 @@ const ComponentInterfaceSymbol EffectOVMusicSeparation::Symbol{ XO("OpenVINO Mus
 
 namespace { BuiltinEffectsModule::Registration< EffectOVMusicSeparation > reg; }
 
-BEGIN_EVENT_TABLE(EffectOVMusicSeparation, wxEvtHandler)
-EVT_CHECKBOX(ID_Type_AdvancedCheckbox, EffectOVMusicSeparation::OnAdvancedCheckboxChanged)
-EVT_BUTTON(ID_Type_DeviceInfoButton, EffectOVMusicSeparation::OnDeviceInfoButtonClicked)
-EVT_BUTTON(ID_Type_ModelManagerButton, EffectOVMusicSeparation::OnModelManagerButtonClicked)
-EVT_CHOICE(ID_Type_ModelSelection, EffectOVMusicSeparation::OnModelSelectionChanged)
-END_EVENT_TABLE()
-
 
 EffectOVMusicSeparation::EffectOVMusicSeparation()
 {
-   try
-   {
-      ov::Core core;
-      auto ov_supported_device = core.get_available_devices();
-      for (auto d : ov_supported_device)
-      {
-         //GNA devices are not supported
-         if (d.find("GNA") != std::string::npos) continue;
 
-         m_simple_to_full_device_map.push_back({ d, core.get_property(d, "FULL_DEVICE_NAME").as<std::string>() });
-
-         mSupportedDevices.push_back(d);
-         mGuiDeviceSelections.push_back({ TranslatableString{ wxString(d), {}} });
-      }
-
-      mGuiSeparationModeSelections.push_back({ TranslatableString{ wxString(" "), {}} });
-      mGuiSeparationModeSelections.push_back({ TranslatableString{ wxString(" "), {}} });
-
-      _populate_model_to_separation_map();
-   }
-   catch (const std::exception& error)
-   {
-      wxLogError("In Music Separation initialization, exception: %s", error.what());
-   }
-}
-
-void EffectOVMusicSeparation::_populate_model_to_separation_map()
-{
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "Drums", "Bass", "Others", "Vocals" };
-      entry.target_stem_for_instrumental = 3; //vocal stem
-      m_model_to_separation_modes.emplace("Demucs v4", entry);
-   }
-
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "Drums", "dummy", "dummy", "dummy" };
-      entry.target_stem_for_instrumental = 0; //drums stem
-      m_model_to_separation_modes.emplace("Demucs v4 FT Drums", entry);
-   }
-
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "dummy", "Bass", "dummy", "dummy" };
-      entry.target_stem_for_instrumental = 1; //bass stem
-      m_model_to_separation_modes.emplace("Demucs v4 FT Bass", entry);
-   }
-
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "dummy", "dummy", "Other Instruments", "dummy" };
-      entry.target_stem_for_instrumental = 2; //others stem
-      m_model_to_separation_modes.emplace("Demucs v4 FT Other Instruments", entry);
-   }
-
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "dummy", "dummy", "dummy", "Vocals" };
-      entry.target_stem_for_instrumental = 3; //vocal stem
-      m_model_to_separation_modes.emplace("Demucs v4 FT Vocals", entry);
-   }
-
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "Drums", "Bass", "Others", "Vocals", "Guitar", "Piano"};
-      entry.target_stem_for_instrumental = 3; //vocal stem
-      m_model_to_separation_modes.emplace("Demucs v4 6s", entry);
-   }
-
-   {
-      SeparationModeEntry entry;
-      entry.stems = { "Vocals" };
-      entry.target_stem_for_instrumental = 0; //vocal stem
-      m_model_to_separation_modes.emplace("MelBandRoformer Vocals (Kimberly Jenson version)", entry);
-   }
-
-   for (auto &pair : m_model_to_separation_modes)
-   {
-      // Count number of non-dummy stems
-      int non_dummy_stems = 0;
-      for (auto &s : pair.second.stems)
-          if (s != "dummy") non_dummy_stems++;
-
-      // Generate the 'all stems' selection string.
-      std::string all_stems_mode = "(" + std::to_string(non_dummy_stems) + " Stem) ";
-      int stems_added = 1;
-      for (auto& s : pair.second.stems)
-      {
-         if (s != "dummy")
-         {
-            all_stems_mode += s;
-            if (stems_added < non_dummy_stems)
-               all_stems_mode += ", ";
-            stems_added++;
-         }
-      }
-
-      if (pair.second.target_stem_for_instrumental >= pair.second.stems.size())
-      {
-         throw std::runtime_error("_populate_model_to_separation_map: pair.second.target_stem_for_instrumental >= pair.second.stems.size");
-      }
-      std::string instrumental_mode = "(2 Stem) " + pair.second.stems[pair.second.target_stem_for_instrumental]
-         + ", Instrumental";
-
-      std::cout << pair.first << ":" << std::endl;
-      std::cout << "  " << all_stems_mode << std::endl;
-      std::cout << "  " << instrumental_mode << std::endl;
-
-      pair.second.guiSeparationModeSelections.push_back({ TranslatableString{ wxString(all_stems_mode), {}} });
-      pair.second.guiSeparationModeSelections.push_back({ TranslatableString{ wxString(instrumental_mode), {}} });
-   }
 }
 
 EffectOVMusicSeparation::~EffectOVMusicSeparation()
@@ -177,672 +59,4 @@ ComponentInterfaceSymbol EffectOVMusicSeparation::GetSymbol() const
 TranslatableString EffectOVMusicSeparation::GetDescription() const
 {
    return XO("Splits a stereo track into 4 new tracks -- Bass, Drums, Vocals, Others");
-}
-
-VendorSymbol EffectOVMusicSeparation::GetVendor() const
-{
-   return XO("OpenVINO AI Effects");
-}
-
-unsigned EffectOVMusicSeparation::GetAudioInCount() const
-{
-   return 2;
-}
-
-// EffectDefinitionInterface implementation
-
-EffectType EffectOVMusicSeparation::GetType() const
-{
-   return EffectTypeProcess;
-}
-
-bool EffectOVMusicSeparation::IsInteractive() const
-{
-   return true;
-}
-
-std::unique_ptr<EffectEditor> EffectOVMusicSeparation::PopulateOrExchange(
-   ShuttleGui& S, EffectInstance&, EffectSettingsAccess&,
-   const EffectOutputs*)
-{
-   mUIParent = S.GetParent();
-
-   auto collection = OVModelManager::instance().GetModelCollection(OVModelManager::MusicSepName());
-   for (auto& model_info : collection->models) {
-      if (model_info->installed) {
-         if (std::find(mSupportedModels.begin(), mSupportedModels.end(), model_info->model_name) == mSupportedModels.end()) {
-            mSupportedModels.push_back(model_info->model_name);
-         }
-      }
-   }
-
-   mGuiModelSelections.clear();
-   for (auto& m : mSupportedModels)
-   {
-      mGuiModelSelections.push_back({ TranslatableString{ wxString(m), {}} });
-   }
-
-   OVModelManager::InstalledCallback callback =
-      [this](const std::string& model_name) {
-      wxTheApp->CallAfter([=]() {
-         if (std::find(mSupportedModels.begin(), mSupportedModels.end(), model_name) == mSupportedModels.end()) {
-            mSupportedModels.push_back(model_name);
-            mGuiModelSelections.push_back({ TranslatableString{ wxString(model_name), {}} });
-         }
-
-         if (mUIParent)
-         {
-            EffectEditor::EnableApply(mUIParent, true);
-            EffectEditor::EnablePreview(mUIParent, false);
-            if (mTypeChoiceModelSelection)
-            {
-               mTypeChoiceModelSelection->Append(wxString(model_name));
-
-               if (mTypeChoiceModelSelection->GetCount() == 1) {
-                  mTypeChoiceModelSelection->SetSelection(mTypeChoiceModelSelection->GetCount() - 1);
-               }
-            }
-         }
-         });
-      };
-
-   OVModelManager::instance().register_installed_callback(OVModelManager::MusicSepName(), callback);
-
-   S.AddSpace(0, 5);
-   S.StartVerticalLay();
-   {
-      S.StartMultiColumn(1, wxLEFT);
-      {
-         auto model_manager_button = S.Id(ID_Type_ModelManagerButton).AddButton(XO("Open Model Manager"));
-      }
-      S.EndMultiColumn();
-
-      S.StartMultiColumn(2, wxLEFT);
-      {
-         mTypeChoiceModelSelection = S.Id(ID_Type_ModelSelection)
-            .MinSize({ -1, -1 })
-            .Validator<wxGenericValidator>(&m_modelSelectionChoice)
-            .AddChoice(XXO("Model Selection:"),
-               Msgids(mGuiModelSelections.data(), mGuiModelSelections.size()));
-      }
-      S.EndMultiColumn();
-
-      S.StartMultiColumn(2, wxLEFT);
-      {
-         //m_deviceSelectionChoice
-         mTypeChoiceSeparationModeCtrl = S.Id(ID_Type)
-            .MinSize({ -1, -1 })
-            .Validator<wxGenericValidator>(&m_separationModeSelectionChoice)
-            .AddChoice(XXO("Separation Mode:"),
-               Msgids(mGuiSeparationModeSelections.data(), mGuiSeparationModeSelections.size()));
-      }
-      S.EndMultiColumn();
-
-      SetModelSeparationModeSelections();
-
-      S.StartStatic(XO(""), wxLEFT);
-      {
-         S.StartMultiColumn(4, wxEXPAND);
-         {
-            mTypeChoiceDeviceCtrl = S.Id(ID_Type)
-               .MinSize({ -1, -1 })
-               .Validator<wxGenericValidator>(&m_deviceSelectionChoice)
-               .AddChoice(XXO("OpenVINO Inference Device:"),
-                  Msgids(mGuiDeviceSelections.data(), mGuiDeviceSelections.size()));
-            S.AddVariableText(XO(""));
-
-            auto device_info_button = S.Id(ID_Type_DeviceInfoButton).AddButton(XO("Device Details..."));
-
-            S.SetStretchyCol(2);
-         }
-         S.EndMultiColumn();
-      }
-      S.EndStatic();
-
-      //advanced options
-      S.StartMultiColumn(2, wxLEFT);
-      {
-         mShowAdvancedOptionsCheckbox = S.Id(ID_Type_AdvancedCheckbox).AddCheckBox(XXO("&Advanced Options"), false);
-      }
-      S.EndMultiColumn();
-
-      S.StartMultiColumn(2, wxLEFT);
-      {
-         mNumberOfShiftsCtrl = S.Name(XO("Shifts"))
-            .Validator<IntegerValidator<int>>(&mNumberOfShifts,
-               NumValidatorStyle::DEFAULT,
-               1,
-               8)
-            .AddTextBox(XO("Shifts"), L"", 12);
-
-         advancedSizer = mNumberOfShiftsCtrl->GetContainingSizer();
-      }
-      S.EndMultiColumn();
-
-   }
-   S.EndVerticalLay();
-
-   show_or_hide_advanced_options();
-
-   return nullptr;
-}
-
-void EffectOVMusicSeparation::show_or_hide_advanced_options()
-{
-   if (advancedSizer)
-   {
-      advancedSizer->ShowItems(mbAdvanced);
-      advancedSizer->Layout();
-   }
-}
-
-void EffectOVMusicSeparation::FitWindowToCorrectSize()
-{
-   if (mUIParent)
-   {
-      mUIParent->Layout();
-      mUIParent->SetMinSize(mUIParent->GetSizer()->GetMinSize());
-      mUIParent->SetSize(mUIParent->GetSizer()->GetMinSize());
-      mUIParent->Fit();
-      auto p = mUIParent->GetParent();
-      if (p)
-      {
-         p->Fit();
-      }
-   }
-}
-
-void EffectOVMusicSeparation::OnAdvancedCheckboxChanged(wxCommandEvent& evt)
-{
-   mbAdvanced = mShowAdvancedOptionsCheckbox->GetValue();
-
-   show_or_hide_advanced_options();
-
-   FitWindowToCorrectSize();
-}
-
-void EffectOVMusicSeparation::OnModelManagerButtonClicked(wxCommandEvent& evt)
-{
-   ShowModelManagerDialog();
-}
-
-void EffectOVMusicSeparation::OnDeviceInfoButtonClicked(wxCommandEvent& evt)
-{
-   std::string device_mapping_str = "";
-   for (auto e : m_simple_to_full_device_map)
-   {
-      device_mapping_str += e.first + " = " + e.second;
-      device_mapping_str += "\n";
-   }
-   auto v = TranslatableString(device_mapping_str, {});
-
-   EffectUIServices::DoMessageBox(*this,
-      v,
-      wxICON_INFORMATION,
-      XO("OpenVINO Device Details"));
-}
-
-void EffectOVMusicSeparation::SetModelSeparationModeSelections()
-{
-   if (!mTypeChoiceSeparationModeCtrl || !mTypeChoiceModelSelection)
-      return;
-
-   int current_selection = mTypeChoiceModelSelection->GetCurrentSelection();
-   if (current_selection == -1)
-   {
-      current_selection = m_modelSelectionChoice;
-   }
-
-   auto selected_model = audacity::ToUTF8(mTypeChoiceModelSelection->GetString(current_selection));
-
-   auto it = m_model_to_separation_modes.find(selected_model);
-   if (it == m_model_to_separation_modes.end())
-   {
-      wxLogError("SetModelSeparationModeSelections: No model entry for selected_model=", selected_model);
-      return;
-   }
-
-   // First, clear the list
-   mTypeChoiceSeparationModeCtrl->Clear();
-
-   for (auto& mode_selection : it->second.guiSeparationModeSelections)
-   {
-      mTypeChoiceSeparationModeCtrl->AppendString(mode_selection.StrippedTranslation());
-   }
-
-   mTypeChoiceSeparationModeCtrl->SetSelection(m_separationModeSelectionChoice);
-
-   FitWindowToCorrectSize();
-}
-
-void EffectOVMusicSeparation::OnModelSelectionChanged(wxCommandEvent& evt)
-{
-   // reset separation mode choice to the first entry in the list.
-   m_separationModeSelectionChoice = 0;
-
-   SetModelSeparationModeSelections();
-}
-
-static std::vector<WaveTrack::Holder> CreateSourceTracks
-(WaveTrack* leader, std::vector<std::string>& labels)
-{
-   std::vector<WaveTrack::Holder> sources;
-   for (auto& label : labels)
-   {
-      WaveTrack::Holder srcTrack = leader->EmptyCopy();
-
-      srcTrack->SetSelected(false);
-      srcTrack->SetSolo(false);
-
-      // append the source name to the track's name
-      srcTrack->SetName(srcTrack->GetName() + wxString("-" + label));
-      sources.emplace_back(srcTrack);
-   }
-   return sources;
-}
-
-static bool HTDemucsProgressUpdate(double perc_complete, void* user)
-{
-   EffectOVMusicSeparation* music_sep = (EffectOVMusicSeparation*)user;
-
-   perc_complete = perc_complete * 100;
-   if (perc_complete > 100)
-      perc_complete = 100;
-
-   return music_sep->UpdateProgress(perc_complete);
-}
-
-bool EffectOVMusicSeparation::UpdateProgress(double perc)
-{
-   //TotalProgress will return true if user clicks 'cancel'
-   if (TotalProgress(perc / 100.0))
-   {
-      std::cout << "User cancelled!" << std::endl;
-      return false;
-   }
-
-   return true;
-}
-
-bool EffectOVMusicSeparation::Process(EffectInstance&, EffectSettings&)
-{
-   try
-   {
-      std::string model_selection_str = mSupportedModels[m_modelSelectionChoice];
-
-      auto model_collection = OVModelManager::instance().GetModelCollection(OVModelManager::MusicSepName());
-
-      // It shouldn't be possible for this condition to be true (User shoudn't have been able to click 'Apply'),
-      // but double check anyway..
-      if (!model_collection || model_collection->models.empty())
-      {
-         throw std::runtime_error("Music Separation models have not been installed.");
-      }
-
-      std::shared_ptr< OVModelManager::ModelInfo > retrieved_model_info;
-      for (auto model_info : model_collection->models) {
-         if (model_info && model_info->installed && model_info->model_name == model_selection_str)
-         {
-            retrieved_model_info = model_info;
-         }
-      }
-
-      if (!retrieved_model_info) {
-         throw std::runtime_error("Couldn't retrieve installed model info for " + model_selection_str);
-      }
-
-      if (!retrieved_model_info->installed) {
-         throw std::runtime_error("This model is not installed: " + retrieved_model_info->model_name);
-      }
-
-      auto model_folder = retrieved_model_info->installation_path;
-
-      std::shared_ptr< ov_demix::DemixModel > model;
-      {
-         auto device = mSupportedDevices[m_deviceSelectionChoice];
-         FilePath cache_folder = FileNames::MkDir(wxFileName(FileNames::DataDir(), wxT("openvino-model-cache")).GetFullPath());
-
-         //Note: Using a variant of wstring conversion that seems to work more reliably when there are special characters present in the path.
-         std::string cache_path = wstring_to_string(wxFileName(cache_folder).GetFullPath().ToStdWstring());
-
-         std::cout << "model_folder = " << model_folder << std::endl;
-         std::cout << "cache_path = " << cache_path << std::endl;
-         std::cout << "number of shifts = " << mNumberOfShifts << std::endl;
-
-         auto create_model_fut = std::async(std::launch::async, [&model_folder, &device, &cache_path, &model_selection_str]()
-         {
-            // WA for OpenVINO locale caching issue (https://github.com/openvinotoolkit/openvino/issues/24370)
-            OVLocaleWorkaround wa;
-
-            std::shared_ptr< ov_demix::DemixModel > ret;
-            if (model_selection_str.find("Demucs") != std::string::npos)
-            {
-               ret = std::make_shared<ov_demix::HTDemucs>(model_folder, device, cache_path);
-            }
-            else if (model_selection_str.find("MelBandRoformer") != std::string::npos)
-            {
-               ret = std::make_shared<ov_demix::MelBandRoformer>(model_folder, device, cache_path, ov_demix::DemixModel::PadMode::Reflect);
-            }
-            else
-            {
-               throw std::runtime_error("Only HTDemucs or MelBandRoformer models are supported right now. ");
-            }
-            return ret;
-
-         });
-
-         std::future_status status;
-         float total_time = 0.f;
-         do {
-            using namespace std::chrono_literals;
-            status = create_model_fut.wait_for(0.5s);
-            {
-               std::string message = "Loading " + model_selection_str + " AI Model to " + device + "...";
-               if (total_time > 10)
-               {
-                  message += " (This could take a while if this is the first time running this feature with this device)";
-               }
-               TotalProgress(0.01, TranslatableString{ wxString(message), {} });
-            }
-
-            total_time += 0.5;
-
-         } while (status != std::future_status::ready);
-
-         model = create_model_fut.get();
-
-         if (!model)
-         {
-            throw std::runtime_error("Error loading model to device...");
-         }
-
-      }
-
-      auto sep_mode_it = m_model_to_separation_modes.find(model_selection_str);
-      if (sep_mode_it == m_model_to_separation_modes.end())
-      {
-         throw std::runtime_error("Did not find separation mode entry for model_selection_str=" + model_selection_str);
-      }
-
-      auto stem_labels = sep_mode_it->second.stems;
-
-      if (m_deviceSelectionChoice >= mSupportedDevices.size())
-      {
-         throw std::runtime_error("Invalid device choice id:  " +
-            std::to_string(m_deviceSelectionChoice));
-      }
-
-      EffectOutputTracks outputs{ *mTracks, GetType(), {{ mT0, mT1 }} };
-
-      bool bGoodResult = true;
-
-      std::cout << "Creating OpenVINO-based HTDemucs object that will run on " << mSupportedDevices[m_deviceSelectionChoice] << std::endl;
-
-      TotalProgress(0.01, XO("Compiling AI Model..."));
-
-      std::vector< WaveTrack::Holder > tracks_to_process;
-      std::vector< int > orig_rates;
-
-      //Create resampled copies of the selected portion of tracks. 
-      // This prevents the 'Resample' operation to modify the user's
-      // original track.
-      for (auto track : outputs.Get().Selected<WaveTrack>())
-      {
-         auto left = track->GetChannel(0);
-         auto start = left->GetStartTime();
-         auto end = left->GetEndTime();
-         if (track->Channels().size() > 1)
-         {
-            auto right = track->GetChannel(1);
-            start = wxMin(start, right->GetStartTime());
-            end = wxMax(end, right->GetEndTime());
-         }
-
-         // Set the current bounds to whichever left marker is
-         // greater and whichever right marker is less:
-         const double curT0 = std::max(start, mT0);
-         const double curT1 = std::min(end, mT1);
-
-         auto start_s = left->TimeToLongSamples(curT0);
-         auto end_s = left->TimeToLongSamples(curT1);
-
-         size_t total_samples = (end_s - start_s).as_size_t();
-         Floats entire_input{ total_samples };
-
-         // create a temporary track list to append samples to
-         auto pTmpTrack = track->EmptyCopy();
-
-         bool bOkay = left->GetFloats(entire_input.get(), start_s, total_samples);
-         if (!bOkay)
-         {
-            throw std::runtime_error("unable to get all left samples. GetFloats() failed for " +
-               std::to_string(total_samples) + " samples");
-         }
-
-         auto iter =
-            pTmpTrack->Channels().begin();
-
-         
-         //append output samples to L & R channels.
-         auto& tmpLeft = **iter++;
-         tmpLeft.Append((samplePtr)entire_input.get(), floatSample, total_samples);
-
-         if (track->Channels().size() > 1)
-         {
-            auto right = track->GetChannel(1);
-            bOkay = right->GetFloats(entire_input.get(), start_s, total_samples);
-            if (!bOkay)
-            {
-               throw std::runtime_error("unable to get all right samples. GetFloats() failed for " +
-                  std::to_string(total_samples) + " samples");
-            }
-
-            auto& tmpRight = **iter;
-            tmpRight.Append((samplePtr)entire_input.get(), floatSample, total_samples);
-         }
-
-         //flush it
-         pTmpTrack->Flush();
-
-         orig_rates.push_back(pTmpTrack->GetRate());
-         if (pTmpTrack->GetRate() != 44100)
-         {
-            pTmpTrack->Resample(44100, mProgress);
-         }
-
-         tracks_to_process.push_back(pTmpTrack);
-      }
-
-      for (size_t ti = 0; ti < tracks_to_process.size(); ti++)
-      {
-         auto pTrack = tracks_to_process[ti];
-
-         auto pLeftChannel = pTrack->GetChannel(0);
-         auto pRightChannel = pLeftChannel;
-         if (pTrack->Channels().size() > 1)
-         {
-            pRightChannel = pTrack->GetChannel(1);
-         }
-
-         double trackStart = pTrack->GetStartTime();
-         double trackEnd = pTrack->GetEndTime();
-
-         // Set the current bounds to whichever left marker is
-         // greater and whichever right marker is less:
-         const double curT0 = trackStart;
-         const double curT1 = trackEnd;
-
-         if (curT1 > curT0) {
-            auto start = pTrack->TimeToLongSamples(curT0);
-            auto end = pTrack->TimeToLongSamples(curT1);
-
-            //Get the length of the buffer (as double). len is
-            //used simple to calculate a progress meter, so it is easier
-            //to make it a double now than it is to do it later
-            auto len = (end - start).as_double();
-
-            size_t total_samples = (end - start).as_size_t();
-
-            auto left_samples = std::make_shared<std::vector<float>>(total_samples);
-            auto right_samples = std::make_shared<std::vector<float>>(total_samples);
-
-            bool bOkay = pLeftChannel->GetFloats(left_samples->data(), start, total_samples);
-            if (!bOkay)
-            {
-               throw std::runtime_error("unable to get all left samples. GetFloats() failed for " +
-                  std::to_string(total_samples) + " samples");
-            }
-
-            bOkay = pRightChannel->GetFloats(right_samples->data(), start, total_samples);
-            if (!bOkay)
-            {
-               throw std::runtime_error("unable to get all right samples. GetFloats() failed for " +
-                  std::to_string(total_samples) + " samples");
-            }
-
-            TotalProgress(0.01, XO("Applying Music Separation using OpenVINO"));
-
-            ov_demix::AudioTrack input_channels = { left_samples, right_samples };
-            auto output_stems = ov_demix::Demix(model, input_channels, mNumberOfShifts, HTDemucsProgressUpdate, this);
-
-            auto pProject = FindProject();
-            const auto& selectedRegion =
-               ViewInfo::Get(*pProject).selectedRegion;
-
-            if (output_stems.size() != stem_labels.size())
-            {
-               throw std::runtime_error("Expected Demix to return " + std::to_string(stem_labels.size()) +
-                  " stems, but instead it returned " + std::to_string(output_stems.size()) + " stems.");
-            }
-
-            // m_separationModeSelectionChoice==1 means instrumental mode is being used.
-            if (m_separationModeSelectionChoice == 1)
-            {
-               auto target_stem_for_instrumental = sep_mode_it->second.target_stem_for_instrumental;
-               if (target_stem_for_instrumental >= output_stems.size())
-               {
-                  throw std::runtime_error("Specified target_stem_for_instrumental out of range");
-               }
-
-               auto target_stem = output_stems[target_stem_for_instrumental];
-               ov_demix::GenerateInstrumental(target_stem, input_channels);
-
-               // replace output stems vector
-               std::vector<ov_demix::AudioTrack> new_output_stems = { target_stem, input_channels};
-               output_stems = new_output_stems;
-
-               // replace stem_labels vector
-               stem_labels = { sep_mode_it->second.stems[target_stem_for_instrumental], "Instrumental" };
-            }
-
-            auto orig_track_name = pTrack->GetName();
-            for (int i = 0; i < stem_labels.size(); i++)
-            {
-               // If it's a dummy stem, skip it.
-               if (stem_labels[i] == "dummy") continue;
-
-               auto &stem_track = output_stems[i];
-
-               // Workaround for 3.4.X issue where setting name of a new output track
-               // retains the label of the track that it was copied from. So, we'll
-               // change the name of the input track here, copy it, and then change it
-               // back later.
-               pTrack->SetName(orig_track_name + wxString(" - " + model_selection_str + " - " + stem_labels[i]));
-
-               //Create new output track from input track.
-               auto newOutputTrack = pTrack->EmptyCopy();
-
-               // create a temporary track list to append samples to
-               auto pTmpTrack = pTrack->EmptyCopy();
-               auto iter = pTmpTrack->Channels().begin();
-
-               //append output samples to L & R channels.
-               auto& tmpLeft = **iter++;
-               tmpLeft.Append((samplePtr)stem_track.first->data(), floatSample, total_samples);
-
-               if (pTrack->Channels().size() > 1)
-               {
-                  auto& tmpRight = **iter;
-                  tmpRight.Append((samplePtr)stem_track.second->data(), floatSample, total_samples);
-               }
-
-               //flush it
-               pTmpTrack->Flush();
-
-               // Clear & Paste into new output track
-               newOutputTrack->ClearAndPaste(selectedRegion.t0() - pTmpTrack->GetStartTime(),
-                  selectedRegion.t1() - pTmpTrack->GetStartTime(), *pTmpTrack);
-
-               //Resample to original tracks rate
-               newOutputTrack->Resample(orig_rates[ti]);
-
-               if (TracksBehaviorsSolo.ReadEnum() == SoloBehaviorSimple)
-               {
-                  //If in 'simple' mode, if original track is solo,
-                  // mute the new track and set it to *not* be solo.
-                  if (newOutputTrack->GetSolo())
-                  {
-                     newOutputTrack->SetMute(true);
-                     newOutputTrack->SetSolo(false);
-                  }
-               }
-
-               // Add the new track list to the output.
-               outputs.AddToOutputTracks(std::move(newOutputTrack));
-
-               // Change name back to original. Part of workaround described above.
-               pTrack->SetName(orig_track_name);
-            }
-         }
-         else
-         {
-             throw std::runtime_error("unexpected case encountered where curT0 (" + std::to_string(curT0) + 
-                ") <= curT1(" + std::to_string(curT1) + ")");
-         }
-      }
-
-      if (bGoodResult)
-         outputs.Commit(); 
-
-      return bGoodResult;
-   }
-   catch (const std::exception& error) {
-      wxLogError("In Music Separation, exception: %s", error.what());
-      EffectUIServices::DoMessageBox(*this,
-         XO("Music Separation failed. See details in Help->Diagnostics->Show Log..."),
-         wxICON_STOP,
-         XO("Error"));
-   }
-
-   return false;
-}
-
-bool EffectOVMusicSeparation::TransferDataToWindow(const EffectSettings&)
-{
-   if (!mUIParent || !mUIParent->TransferDataToWindow())
-   {
-      return false;
-   }
-
-   OVModelManager::InstalledCallback callback =
-      [this](const std::string& model_name) {
-      wxTheApp->CallAfter([=]() {
-         EffectEditor::EnableApply(mUIParent, true);
-         EffectEditor::EnablePreview(mUIParent, false);
-         });
-      };
-   OVModelManager::instance().register_installed_callback(OVModelManager::MusicSepName(), callback);
-
-   auto model_collection = OVModelManager::instance().GetModelCollection(OVModelManager::MusicSepName());
-   if (!model_collection || model_collection->models.empty() || !model_collection->models[0]->installed)
-   {
-      EffectEditor::EnableApply(mUIParent, false);
-   }
-   else
-   {
-      EffectEditor::EnableApply(mUIParent, true);
-   }
-   EffectEditor::EnablePreview(mUIParent, false);
-
-   return true;
 }
